requirements.txt
streamlit
pandas
numpy
cohere
openai
faiss-cpu
sentence-transformers
python-dotenv
scikit-learn